~ move datatype descriptions and definitions out of "future" package
