+++ opc-ua-stack/bsd-parser/bsd-parser-core/src/main/java/org/eclipse/milo/opcua/binaryschema/generator/BinaryDataTypeDictionaryGenerator.java
@@ -25,12 +25,12 @@ import javax.xml.namespace.QName;
 import com.sun.xml.bind.marshaller.NamespacePrefixMapper;
 import org.eclipse.milo.opcua.stack.core.types.builtin.NodeId;
 import org.eclipse.milo.opcua.stack.core.types.builtin.QualifiedName;
-import org.eclipse.milo.opcua.stack.core.types.structured.future.EnumDefinition;
-import org.eclipse.milo.opcua.stack.core.types.structured.future.EnumDescription;
-import org.eclipse.milo.opcua.stack.core.types.structured.future.EnumField;
-import org.eclipse.milo.opcua.stack.core.types.structured.future.StructureDefinition;
-import org.eclipse.milo.opcua.stack.core.types.structured.future.StructureDescription;
-import org.eclipse.milo.opcua.stack.core.types.structured.future.StructureField;
+import org.eclipse.milo.opcua.stack.core.types.structured.EnumDefinition;
+import org.eclipse.milo.opcua.stack.core.types.structured.EnumDescription;
+import org.eclipse.milo.opcua.stack.core.types.structured.EnumField;
+import org.eclipse.milo.opcua.stack.core.types.structured.StructureDefinition;
+import org.eclipse.milo.opcua.stack.core.types.structured.StructureDescription;
+import org.eclipse.milo.opcua.stack.core.types.structured.StructureField;
 import org.eclipse.milo.opcua.stack.core.util.Namespaces;
 import org.opcfoundation.opcua.binaryschema.ByteOrder;
 import org.opcfoundation.opcua.binaryschema.EnumeratedType;
